Refactor persistence : more readable API for webhook payload
