+++ specServer/webhook_spec.js
@@ -53,103 +53,3 @@ describe("The webhook",function(){
   });
 });
 
-describe("Payload extractor", function() {
-
-  var extractor = require('../persistence/webhook-input');
-
-  var payload;
-
-  beforeEach(function() {
-    payload = { 
-      ref: "refs/heads/whatever",
-      commits: [
-        { added: [], removed: [], modified: [] },
-        { added: [], removed: [], modified: [] },
-        { added: [], removed: [], modified: [] },
-      ],
-      head_commit: { id: "sha123" },
-    }
-  });
-
-  it("extracts branch name", function() {
-    expect(extractor.branch(payload)).toEqual("whatever");
-  });
-
-  it("extracts last commit id",function() {
-    expect(extractor.headCommitId(payload)).toEqual("sha123");
-  });
-
-  it("extracts every altered file",function(){
-    payload.commits[0].added=    [ "addedFile" ];
-    payload.commits[0].modified= [ "modifiedFile" ];
-    payload.commits[0].removed=  [ "removedFile" ];
-
-    expect(extractor.extractAndDigest(payload)).toEqual(
-      {
-         added: [ "addedFile" ],
-         modified: [ "modifiedFile" ],
-         removed: [ "removedFile" ]
-      }
-    )
-
-  });
-
-  it("digest file modified in serveral commits, but only once per file",function(){
-    payload.commits[0].modified= [ "file", "anotherFile" ];
-    payload.commits[1].modified= [ "file" ]
-
-    expect(extractor.extractAndDigest(payload).modified).toEqual(jasmine.arrayContaining(
-          ["file", "anotherFile"]));
-  });
-
-  it("digest as removed file modified then removed",function(){
-    payload.commits[0].modified = [ "file" ];
-    payload.commits[1].removed = [ "file" ];
-
-    expect(extractor.extractAndDigest(payload)).toEqual(
-        {added:[],modified:[],removed:["file"]});
-  });
-
-  it("ignores files created first and removed last",function(){
-    payload.commits[0].added = [ "file" ];
-    payload.commits[1].modified = [ "file" ];
-    payload.commits[2].removed = [ "file" ];
-
-    expect(extractor.extractAndDigest(payload)).toEqual(
-        {added:[],modified:[],removed:[]});
-  });
-
-  it("digest as modified a file removed then added",function(){
-    payload.commits[0].removed = [ "file" ];
-    payload.commits[1].added = [ "file" ];
-    expect(extractor.extractAndDigest(payload)).toEqual(
-        {added:[],modified:["file"],removed:[]});
-  });
-
-  it("digest as modified a file modified, then removed, then added",function(){
-    payload.commits[0].modified = [ "file" ];
-    payload.commits[1].removed = [ "file" ];
-    payload.commits[2].added = [ "file" ];
-
-    expect(extractor.extractAndDigest(payload)).toEqual(
-        {added:[],modified:["file"],removed:[]});
-  });
-
-  it("digest as added a file added, then not removed",function(){
-    payload.commits[0].added = [ "file" ];
-    payload.commits[1].modified = [ "file" ];
-
-    expect(extractor.extractAndDigest(payload)).toEqual(
-        {added:["file"],modified:[],removed:[]});
-  });
-});
-
-/*
-digest :
- added, ... , removed -> ignored
- ... removed -> removed
- modified. ... , removed, added -> modified
- removed, ...  added -> modified
- ..., any -> any
- */
-
